+++ api/src/main/java/vook/server/api/web/auth/app/JWTHelperProvider.java
@@ -1,8 +1,10 @@
-package vook.server.api.helper.jwt;
+package vook.server.api.web.auth.app;
 
 import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+import vook.server.api.helper.jwt.JWTReader;
+import vook.server.api.helper.jwt.JWTWriter;
 
 @Component
 public class JWTHelperProvider {
